+++ vonage/CHANGES.md
@@ -1,5 +1,10 @@
-# 3.99.2a0
+# 3.99.3a0
 - Add support for the [Vonage Subaccounts API](https://developer.vonage.com/en/use-cases/using-subaccounts).
+
+# 3.99.2a0
+- Add support for the [RCS messaging channel](https://developer.vonage.com/en/messages/concepts/rcs) of the Vonage Messages API
+- Add method to revoke an RCS message
+- Add method to mark a WhatsApp message as read
 
 # 3.99.0a12
 - Add support for the [Vonage Account API](https://developer.vonage.com/en/account/overview).